+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Release 1.0.1
 
-* [Refactor]: removed public router() method from App class
-* [Refactor]: refactored app() function
+* **[Feat]**: added support to disable Gutenberg Editor
+* **[Refactor]**: removed public router() method from App class
+* **[Refactor]**: refactored app() function
+
+Release Date: August 17, 2021
 
 ## Core Release